Rename 'irls.cpp' -> 'robust.cpp'
