+++ localization/include/lidar_feature_localization/optimizer.hpp
@@ -35,7 +35,7 @@
 #include <tuple>
 #include <vector>
 
-#include "lidar_feature_localization/irls.hpp"
+#include "lidar_feature_localization/robust.hpp"
 #include "lidar_feature_localization/matrix_type.hpp"
 #include "lidar_feature_localization/math.hpp"
 #include "lidar_feature_localization/posevec.hpp"
