+++ localization/test/test_robust.cpp
@@ -26,14 +26,74 @@
 // ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 
-#ifndef LIDAR_FEATURE_LOCALIZATION__IRLS_HPP_
-#define LIDAR_FEATURE_LOCALIZATION__IRLS_HPP_
 
-#include <Eigen/Core>
+#include <gtest/gtest.h>
 
-double MedianAbsoluteDeviation(const Eigen::VectorXd & v);
-double Scale(const Eigen::VectorXd & v);
-double Huber(const double e, const double k = 1.345);
-double HuberDerivative(const double e, const double k = 1.345);
+#include <cmath>
+#include <random>
 
-#endif  // LIDAR_FEATURE_LOCALIZATION__IRLS_HPP_
+#include "lidar_feature_localization/robust.hpp"
+
+TEST(IRLS, MedianAbsoluteDeviation)
+{
+  {
+    Eigen::VectorXd v(5);
+    v << 7, 9, 3, 0, 1;
+
+    // median(v) = 3
+    // | v - median(v) | = [4, 6, 0, 3, 2]
+    // median(| v - median(v) |) = 3
+
+    EXPECT_EQ(MedianAbsoluteDeviation(v), 3);
+  }
+
+  {
+    Eigen::VectorXd v(6);
+    v << 8, 3, 4, 0, 5, 1;
+
+    // median(v) = 3.5
+    // | v - median(v) | = [4.5, 0.5, 0.5, 3.5, 1.5, 2.5]
+    // median(| v - median(v) |) = (1.5 + 2.5) / 2 = 2.0
+
+    EXPECT_EQ(MedianAbsoluteDeviation(v), 2.);
+  }
+}
+
+TEST(Scale, StandardDeviation)
+{
+  // generate random numbers, compute the sample variance, and compare to the
+  // sample variance
+
+  const double mean = 0.;
+  const double stddev = 10.;
+  const int n = 100000;
+
+  const Eigen::VectorXd errors = [&] {
+      std::random_device rd{};
+      std::mt19937 gen{rd()};
+
+      std::normal_distribution<> normal{mean, stddev};
+
+      Eigen::VectorXd errors(n);
+      for (int i = 0; i < n; ++i) {
+        errors(i) = normal(gen);
+      }
+      return errors;
+    }();
+
+  const double nf = static_cast<double>(n);
+  const double sample_stddev = std::sqrt(((nf - 1) / nf) * (stddev * stddev));
+  const double scale = Scale(errors);
+
+  EXPECT_LE(std::fabs(sample_stddev - scale), 0.05);
+}
+
+TEST(HuberDerivative, NumericalDiff)
+{
+  const double k = 1.0;
+
+  {
+    const double d = (Huber(0.90 + 1e-4, k) - Huber(0.90, k)) / 1e-4;
+    ASSERT_LT(std::fabs(d - HuberDerivative(0.90)), 1e-3);
+  }
+}
